csum: egress: support skb->ip_summed == CHECKSUM_PARTIAL
When testing with real HW, we could noticed that the packet was modified
after the TC egress hook: the L4 checksum was modified, but at the wrong
place because skb->csum_offset was pointing to the TCP checksum offset,
and not the UDP one after the modification done in eBPF. The packet was
then corrupted, with a wrong UDP checksum only covering the pseudo
header.

One solution could be to add a new eBPF helper fixing skb->csum_offset.
By doing that, the checksum could be correctly computed in HW, at the
right spot.

Instead, as a workaround, we can use TC csum action, which will
recompute the csum, and set skb->ip_summed = CHECKSUM_NONE. To be able
to use it, the eBPF hook needs to return TC_ACT_PIPE instead of _OK. As
a result, the L4 checksum will no longer be corrupted after us.

To get some stats to check that everything is working correctly, 'tc -s
action show action csum' can be used (eventually with '-j' for JSON
support).

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>