test: disable gro/gso on client/server
We cannot just set gso_max_segs on the hosts doing the modification,
because in this particular setup, they will receive aggregated packets,
and are not the initiator of the original packets.

Instead, we could set gso_max_segs on the client and server, but it will
slightly affect the perf in this test lab.

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>