ingress: support non-linear skb
With some NICs, we could get only access to the UDP header size, not the
whole packet.

In this case, we will need to pull more data. Use bpf_skb_load_bytes()
for that. Then it is required to use bpf_skb_store_bytes().

Use BPF_F_RECOMPUTE_CSUM when fields are actually modified.

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>