tc: fix misaligned packet access error thrown by BPF verifier
The ipv6hdr and iphdr variables in tc_tcp_in_udp() are modified multiple
times. The verifier doesn't seem to understand that one of these two
modifications is under an if block and the other is under the affine else
block so it rejects running the programme due to misaligned packet access
when attempting to load the programme on certain interfaces such as a DSA
user interface. Declare new variables for the else block to satisfy the
verifier.

Fixes: 58342d35a6f1 ("tc: support layer 3 interfaces")
Signed-off-by: Chester A. Unal <chester.a.unal@arinc9.com>